refactor: WaitForDeployment works with deployment
